Inline async action definition with implementation
This is similar to the previous commit (combining implementation and
type signatures together), but, you know, for async action creators
instead.

That all happened surprisingly easily.
