+++ src/actions/create-async-action.ts
@@ -1,7 +1,38 @@
-import { ActionConstant } from '../types/actions';
 import Phase from '../constants/phase';
 import bindActionType from './bind-action-type';
 import validateActionType from './validate-action-type';
+import {
+  ActionConstant,
+  CoercibleAction,
+  ActionSuccess,
+  ActionFailure,
+  OptimisticAction,
+} from '../types/actions';
+
+/**
+ * Wraps an async function in an action creator. It optimistically
+ * dispatches immediately, then again when the promise resolves.
+ * @param type A unique name which describes the action.
+ * @param effect Any function that returns a promise.
+ *
+ * @example
+ * createAction.async('account/save-settings', (settings: AccountSettings) => {
+ *   await http.put(`/accounts/${accountId}/settings/`, settings)
+ * })
+ */
+export default function createAsyncAction<Effect extends AsyncFunction>(
+  type: ActionConstant,
+  effect: Effect,
+): Effect extends () => Promise<infer TReturn>
+  ? CoercibleAction<[], ActionSequence<void, TReturn>>
+  : never;
+
+export default function createAsyncAction<Effect extends AnyAsyncFunction>(
+  type: ActionConstant,
+  effect: Effect,
+): Effect extends (input: infer Input, ...args: any) => Promise<infer TReturn>
+  ? CoercibleAction<[Input], ActionSequence<Input, TReturn>>
+  : never;
 
 export default function createAsyncAction<
   TReturn,
@@ -41,3 +72,14 @@ export default function createAsyncAction<
 
   return bindActionType(actionType, createAsyncAction);
 }
+
+type AsyncFunction = () => Promise<any>;
+type AnyAsyncFunction = (...args: any) => Promise<any>;
+
+type ActionSequence<Optimistic, TReturn> = AsyncGenerator<
+  | OptimisticAction<Optimistic>
+  | ActionSuccess<TReturn>
+  | ActionFailure<unknown>,
+  TReturn,
+  void
+>;
